match_sf_clusters(): fix failure to drop non-overlapping intervals
testing showed logic would fully match 2 clusters separated by more days than the value specified for `days_thresh`. The reason was the `| !is.na(VALID_TIME)` side of the condition, which makes  the condition to always return `TRUE`, i.e. to always keep the row regardless of the value of `VALID_TIME`