Added `format_iso8601()`
This function is attempting to solve error occurring when using the App in MoveApps (worked fine locally), by explicitly definint the date-times as ISO8601 with a colon in the timezone component (a requirement of he extended version of the ISO8601, which apparently ER expects for radio agent POST ??). Anyway, fingers-crossed it makes the trick.